src/event/requirements.txt
PyNaCl
gspread
